♻️ 필요없는 코드 삭제, FocusViewControllerFactory 분리
