--- JipJung/JipJung/UI/Home/Focus/FocusExtraView.swift
@@ -1,15 +1,0 @@
-//
-//  FocusExtraView.swift
-//  JipJung
-//
-//  Created by 윤상진 on 2021/11/07.
-//
-
-import UIKit
-
-class FocusExtraView: UIView {
-    override func draw(_ rect: CGRect) {
-        super.draw(rect)
-        backgroundColor = .systemCyan
-    }
-}
