--- JipJung/JipJung/UI/Home/Focus/FocusStartViewProtocol.swift
@@ -1,19 +1,0 @@
-//
-//  FocusStartViewProtocol.swift
-//  JipJung
-//
-//  Created by 윤상진 on 2021/11/08.
-//
-
-import UIKit
-
-protocol FocusStarViewDelegate: AnyObject {
-    func startButtonDidClicked()
-}
-
-protocol FocusStartViewable: UIView {
-    var delegate: FocusStarViewDelegate? { get set }
-    func clickStartButton(_ sender: UIButton)
-    func startFocus()
-    func endFocus()
-}
